Dynamically recompile the plugin when the availability of EQC changes
Since EQC is a third party library that is usually installed into the
system path, often it's availability changes as the erlang version
changes or EQC is installed. The plugin adapts to these changes by
recompiling itself to match the current EQC status for the current
Erlang install and thus avoids having to manually clean up the plugin to
force it to recompile.